# Optimize ninjaString.ValueWithEscaper